fix(inventory-change-value): having debounce for changing values
