+++ src/locales/de-DE/messages.po
@@ -376,7 +376,7 @@ msgstr "Alarmierungsintegrationen (PD, Slack, Discord, Teams)"
 msgid "Alerting Settings"
 msgstr "Benachrichtigungseinstellungen"
 
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:123
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:142
 msgid "All"
 msgstr ""
 
@@ -615,8 +615,8 @@ msgstr ""
 msgid "Change Product Tier"
 msgstr "Produktstufe ändern"
 
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:86
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:106
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:105
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:125
 msgid "Change Type"
 msgstr ""
 
@@ -645,7 +645,7 @@ msgstr ""
 msgid "Check the newly created role assignment"
 msgstr ""
 
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:147
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:161
 msgid "Clear"
 msgstr ""
 
@@ -738,7 +738,7 @@ msgstr "Gemeinschaftliche Unterstützung"
 msgid "Compliance"
 msgstr "Compliance"
 
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:31
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:25
 msgid "Compliant"
 msgstr ""
 
@@ -869,7 +869,7 @@ msgstr ""
 msgid "Create Service Account"
 msgstr ""
 
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:35
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:29
 #: src/pages/panel/user-settings/UserSettingsApiTokens.tsx:27
 msgid "Created"
 msgstr ""
@@ -980,7 +980,7 @@ msgstr "Löschen"
 msgid "Delete Provider"
 msgstr "Anbieter löschen"
 
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:39
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:33
 msgid "Deleted"
 msgstr ""
 
@@ -2074,9 +2074,9 @@ msgstr "Umbenennen"
 msgid "Reset Password"
 msgstr "Passwort zurücksetzen"
 
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:35
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:37
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:39
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:29
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:31
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:33
 msgid "Resource"
 msgstr ""
 
@@ -2169,8 +2169,8 @@ msgstr "Sicherer Lesezugriff"
 msgid "Secure Your Cloud, Elevate Your Peace of Mind"
 msgstr ""
 
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:31
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:33
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:25
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:27
 #: src/pages/panel/inventory/InventoryTemplateBoxes.tsx:33
 #: src/shared/layouts/panel-layout/menuList.tsx:62
 msgid "Security"
@@ -2211,7 +2211,7 @@ msgstr "Sicherheitsbewertung"
 msgid "See this Resource in AWS Console"
 msgstr "Sehen Sie sich diese Ressource in der AWS-Konsole an"
 
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:147
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:166
 msgid "Select All"
 msgstr ""
 
@@ -2544,7 +2544,7 @@ msgstr "E-Mail aktualisieren"
 msgid "Update Password"
 msgstr "Kennwort aktualisieren"
 
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:37
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:31
 msgid "Updated"
 msgstr ""
 
@@ -2593,7 +2593,7 @@ msgstr "Wert"
 msgid "Version"
 msgstr ""
 
-#: src/pages/panel/inventory/inventory-form/InventoryFormChangesValue.tsx:33
+#: src/pages/panel/inventory/inventory-form/InventoryFormChangePicker.tsx:27
 msgid "Vulnerable"
 msgstr ""
